pyproject.toml
[project]
name = "youtube-digest"
version = "1.0.0"
description = "Automated YouTube video summarizer with email delivery"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "youtube-transcript-api>=1.0.0",
    "feedparser>=6.0.0",
    "requests>=2.28.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "markdown>=3.5.0",
    "pyyaml>=6.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "apscheduler>=3.10.0",
]

[project.scripts]
youtube-digest = "youtube_digest.__main__:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["youtube_digest*"]
